Fix RoPE precision: rotation in f32, sync with upstream PT
RoPE rotation (x*cos + rotate_half(x)*sin) was happening in bfloat16
when model weights were bf16, causing silent precision loss — same bug
fixed in CanViT-PyTorch@217f156.

Fix: compute_rope always returns f32, apply_rope_with_prefix upcasts x
to f32 for rotation then casts back. Verified bitwise: bf16 output
matches f32→bf16 (diff=0.0).

Also syncs with upstream PT changes:
- n_register_tokens → n_backbone_registers (registers moved to model)
- Add enable_reads config flag for canvas read ablation
- Update convert.py config extraction, tests for new PT APIs
- Fix stale weights path in push_to_hub.py
- Add Demos section to README, fix quickstart snippet

+++ tests/unit/rope/test_rope.py
@@ -38,7 +38,7 @@ class TestSinCos:
 
         for suffix, hd in [("bb", 64), ("ca", 128)]:
             periods = pt_make_rope_periods(head_dim=hd, base=100.0, device=torch.device("cpu"))
-            rope = pt_compute_rope(positions=local_pos, periods=periods, dtype=torch.float32)
+            rope = pt_compute_rope(positions=local_pos, periods=periods)
 
             pos_mlx = mx.array(grid_coords(8, 8).reshape(1, 64, 2))
             sin, cos = compute_rope(pos_mlx, make_rope_periods(hd))
